refactor(agent): keep normalized publish identity behind the internal boundary
