+++ packages/agent/scripts/test-package-root.mjs
@@ -17,6 +17,19 @@ const publicCatalogActivation = await import(
 const registeredAuthorityContract = await import(
   '@origintrail-official/dkg-agent/dist/registered-context-graph-authority.js'
 );
+const legacyAuthorResolver = await import(
+  '@origintrail-official/dkg-agent/dist/finalized-assertion-author.js'
+);
+const publishSelectionContract = await import(
+  '@origintrail-official/dkg-agent/dist/publish-author-selection.js'
+);
+if (typeof legacyAuthorResolver.resolveFinalizedAssertionAuthor !== 'function'
+  || Object.keys(legacyAuthorResolver).join(',') !== 'resolveFinalizedAssertionAuthor') {
+  throw new Error('the released author resolver must expose only its public compatibility function');
+}
+if (Object.keys(publishSelectionContract).length !== 0) {
+  throw new Error('public publish selection contracts must not expose internal normalization');
+}
 const require = createRequire(import.meta.url);
 const packageManifest = require('@origintrail-official/dkg-agent/package.json');
 const packageExports = packageManifest.exports;
@@ -79,22 +92,25 @@ for (const [subpath, target] of Object.entries(packageExports)) {
     throw new Error(`internal export exception must remain blocked: ${subpath}`);
   }
 }
-const representativeInternalSpecifier =
-  '@origintrail-official/dkg-agent/dist/internal/context-graph-authority/' +
-  'context-graph-agent-gate-authority.js';
-try {
-  await import(representativeInternalSpecifier);
-  throw new Error(`internal module unexpectedly resolved: ${representativeInternalSpecifier}`);
-} catch (error) {
-  if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
-}
-try {
-  require.resolve(representativeInternalSpecifier);
-  throw new Error(
-    `internal module unexpectedly resolved via require: ${representativeInternalSpecifier}`,
-  );
-} catch (error) {
-  if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+for (const path of [
+  'context-graph-authority/context-graph-agent-gate-authority.js',
+  'resident-assertion-author-selection.js',
+  'publish-identity-plan.js',
+  'finalized-assertion-author.js',
+]) {
+  const specifier = `@origintrail-official/dkg-agent/dist/internal/${path}`;
+  try {
+    await import(specifier);
+    throw new Error(`internal module unexpectedly resolved: ${specifier}`);
+  } catch (error) {
+    if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+  }
+  try {
+    require.resolve(specifier);
+    throw new Error(`internal module unexpectedly resolved via require: ${specifier}`);
+  } catch (error) {
+    if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+  }
 }
 const legacySynchronizationError = new legacyCatalogSync.Rfc64CatalogSynchronizationErrorV1(
   'no-authorized-provider',
